Code analysis bot does not need to comment on merge commits
Summary:
This PR fixes a CI issue that came up on #11304 after merging master into the pr branch.

That created a merge commit with over 1000 changed files. That might have created irrelevant comments, but the github commit api is limited to 300 files, with only the first 294 having the expected diffs included.

I can't think of any reasonable scenario where a merge would have changes that aren't either on master or another commit in the PR, so I've updated it to treat merge commits as having no changes.
Closes https://github.com/facebook/react-native/pull/13260

Differential Revision: D4819805

Pulled By: ericvicenti

fbshipit-source-id: 7b6b51b18bb36503a719a3ba5c9163cd6ef00e17